move `Vec3` and `Mat3` into `types`
