+++ src/octree/bbox.rs
@@ -1,5 +1,4 @@
-use crate::math::min_and_max;
-use crate::vec3::Vec3;
+use crate::{math::min_and_max, types::Vec3};
 
 /// Determines the location and extent of a collection of source points
 #[derive(Debug, Clone, Copy)]
