+++ src/sources/point.rs
@@ -1,4 +1,4 @@
-use crate::vec3::Vec3;
+use crate::types::Vec3;
 use std::f64::consts::PI;
 
 /// H-field generated by a infinitesimal point current source
