--- src/vec3.rs
@@ -1,94 +1,0 @@
-//! A 3-length vector
-
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-#[repr(transparent)]
-pub struct Vec3(pub [f64; 3]);
-
-impl Vec3 {
-    /// Construct a new Vec3 from a tuple that contains slices, at index `idx`
-    pub fn from_slice_tuple(slices: (&[f64], &[f64], &[f64]), idx: usize) -> Self {
-        Vec3([slices.0[idx], slices.1[idx], slices.2[idx]])
-    }
-
-    /// Cross-product that returns a new Vec3
-    pub fn cross(&self, other: &Vec3) -> Vec3 {
-        Vec3([
-            self.0[1] * other.0[2] - self.0[2] * other.0[1],
-            self.0[2] * other.0[0] - self.0[0] * other.0[2],
-            self.0[0] * other.0[1] - self.0[1] * other.0[0],
-        ])
-    }
-
-    pub fn dot(&self, other: &Vec3) -> f64 {
-        self[0] * other[0] + self[1] * other[1] + self[2] * other[2]
-    }
-
-    /// Vector magnitude
-    pub fn mag(&self) -> f64 {
-        self.dot(self).sqrt()
-    }
-
-    pub fn to_slice(&self) -> &[f64; 3] {
-        &self.0
-    }
-}
-
-impl std::ops::Index<usize> for Vec3 {
-    type Output = f64;
-
-    fn index(&self, index: usize) -> &Self::Output {
-        &self.0[index]
-    }
-}
-
-impl std::ops::IndexMut<usize> for Vec3 {
-    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-        &mut self.0[index]
-    }
-}
-
-impl std::ops::AddAssign for Vec3 {
-    fn add_assign(&mut self, rhs: Self) {
-        self.0[0] += rhs.0[0];
-        self.0[1] += rhs.0[1];
-        self.0[2] += rhs.0[2];
-    }
-}
-
-impl std::ops::Add for Vec3 {
-    type Output = Vec3;
-    fn add(self, rhs: Self) -> Self::Output {
-        Self([
-            self.0[0] + rhs.0[0],
-            self.0[1] + rhs.0[1],
-            self.0[2] + rhs.0[2],
-        ])
-    }
-}
-
-impl std::ops::Sub for Vec3 {
-    type Output = Vec3;
-    fn sub(self, rhs: Self) -> Self::Output {
-        Self([
-            self.0[0] - rhs.0[0],
-            self.0[1] - rhs.0[1],
-            self.0[2] - rhs.0[2],
-        ])
-    }
-}
-
-// allow scaling
-impl std::ops::Mul<f64> for Vec3 {
-    type Output = Vec3;
-    fn mul(self, rhs: f64) -> Self::Output {
-        Self([self.0[0] * rhs, self.0[1] * rhs, self.0[2] * rhs])
-    }
-}
-
-impl std::ops::MulAssign<f64> for Vec3 {
-    fn mul_assign(&mut self, rhs: f64) {
-        self[0] *= rhs;
-        self[1] *= rhs;
-        self[2] *= rhs;
-    }
-}
